Add non-blocking write functions to GkrellmdClient
Add infrastructure for doing non-blocking writes to connected clients.
All to-be-sent bytes are first pushed to a write-buffer which is then
written to the output stream unless the write-operation would block.
Instead of blocking, we setup a notification GSource that tells us when
to retry the write operation.

For further robustness this code should probably add some upper limit on
the length of the write-buffer.